Create nowherezeromatrixconstruction
This code in SAGE follows the steps in Theorem ?? of "The combinatorial inverse eigenvalue problems: complete graphs and small graphs with strict inequality, W. Barrett, A. Lazenby, N. Malloy, C. Nelson, W. Sexton, Electronic Journal of Linear Algebra, 656-672 (26) 2013" to construct a nowherezero matrix for given eigenvalues. The necessary and sufficient conditions for this matrix to exist is that the list L contains at least two distinct numbers. The codes checks this and prints an error if it is not satisfied.